fix: move the overlay handling to the browser
